snd_tascam: add object for sound unit of TASCAM FireWire series
In TASCAM FireWire models, FW1884, FW1804, and FW1082 transmits
isochronous packets with PCM frames as well as image of device state.
ALSA firewire-tascam driver has parser of the image and allows userspace
application to retrieve it.

This commit adds GObject-derived object for the function. The object
implements Hitaki.AlsaFirewire and Hitaki.TascamProtocol.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>