snd_motu_register_dsp_parameter: add boxed object for parameters of register DSP in MOTU FireWire series
In MOTU FireWire series, the sound unit transmits isochronous packets
with PCM frames and MIDI messages as well as DSP parameters. ALSA
firewire-motu driver implements parser for the parameters since Linux
kernel v5.16 or later, and allows userspace application to retrieve the
parameters in a shape of structure.

This commit adds Boxed object for the structure.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>