change licencing to LGPL-2.1-or-later
It appears that it is not permitted to put GPL-2.0-only code into
LGPL-3.0-or-later code.

https://www.gnu.org/licenses/gpl-faq.en.html#AllCompatibility

Unfortunately, one of UAPI headers of Linux kernel
(include/sound/firewire.h) is included in libhitaki for backporting,
and it is licensed under GPL-2.0-only. The last change of the header is
for v6.3 kernel, therefore it seems to be preferable to keep it for a
while.

It is granted to put GPL-2.0-only code into LGPL-2.1-or-later code. This
commit relicenses libhitaki itself under LGPL-2.1-or-later.

Signed-off-by: Takashi Sakamoto <o-takashi@sakamocchi.jp>